pyproject.toml
[project]
name = "artrestorer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.3",
    "google-cloud-storage>=3.6.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "uvicorn>=0.38.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
]
